refactored name of apiSpecSample.json
